+++ app/scripts/Track.js
@@ -1,10 +1,17 @@
 import { scaleLinear } from 'd3-scale';
+import { fake as fakePubSub } from './hocs/with-pub-sub';
 
 // Services
 import { isWithin } from './utils';
 
 class Track {
   constructor({ id, pubSub, getTheme }) {
+    if (pubSub) {
+      this.pubSub = pubSub;
+    } else {
+      this.pubSub = fakePubSub;
+    }
+
     this.id = id;
     this._xScale = scaleLinear();
     this._yScale = scaleLinear();
@@ -19,9 +26,13 @@ class Track {
     this.position = [0, 0];
     this.dimensions = [1, 1];
     this.options = {};
-    this.pubSub = pubSub;
     this.pubSubs = [];
-    this.getTheme = getTheme;
+
+    if (getTheme) {
+      this.getTheme = getTheme;
+    } else {
+      this.getTheme = () => {};
+    }
 
     this.pubSubs.push(
       this.pubSub.subscribe(