+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Add infrastructure for value scale zooming
 - Fixed the colors when dragging a track over a higlass display
 - Added support for value scale zooming
+- Added utils/track-utils to provide track functionality for use without inheritance
 
 _[Detailed changes since v1.6.7](https://github.com/higlass/higlass/compare/v1.6.7...develop)_